+++ tests/testthat/test_efficiency_frontier.R
@@ -14,8 +14,8 @@ test_that(
     
     result1 <- get_frontier(test1)
     
-    expect_identical(
-      sort(sort(result1)),
+    expect_equivalent(
+      result1,
       c("Scenario 1", "Scenario 5", "Scenario 8")
     )  
     
@@ -30,8 +30,8 @@ test_that(
       .effect=c(9.13,9.635,11.1654,11.5152,11.582), stringsAsFactors=FALSE)
     
     result2 <- get_frontier(test2)
-    expect_identical(
-      sort(result2),
+    expect_equivalent(
+      result2,
       c("Scenario 2", "Scenario 5")
     )
     
@@ -44,9 +44,8 @@ test_that(
                       .effect=c(10.2497,11.1654,11.3,11.6588,11.7,12.1), stringsAsFactors=FALSE)
     
     result3 <- get_frontier(test3)
-    
-    expect_identical(
-      sort(result3),
+    expect_equivalent(
+      result3,
       c("Scenario 1", "Scenario 2", "Scenario 4", "Scenario 6")
     )
     
@@ -60,8 +59,8 @@ test_that(
                         .effect=c(10.2497,11.1654,11.5152,11.6248,11.6296,11.6296), stringsAsFactors=FALSE)
     
     result4 <- get_frontier(test4)    
-    expect_identical(
-      sort(result4),
+    expect_equivalent(
+      result4,
       c("Scenario 1", "Scenario 6")
     )
     
@@ -74,8 +73,8 @@ test_that(
                       .effect=c(10.2497,11.1654,11.6296,11.7,11.8), stringsAsFactors=FALSE)
     
     result5 <- get_frontier(test5)
-    expect_identical(
-      sort(result5),
+    expect_equivalent(
+      result5,
       c("Scenario 1", "Scenario 2", "Scenario 3", "Scenario 4", "Scenario 5")
     )
     
@@ -87,12 +86,11 @@ test_that(
     test6<-data.frame(.strategy_names=c("Scenario 1","Scenario 2","Scenario 3","Scenario 4","Scenario 5","Scenario 6","Scenario 7"),
                       .cost=c(25000,35000,50000,54651,58500,64266,91744),
                       .effect=c(10.2497,10.2497,11.3,11.6588,11.6588,11.7,12.1),
-                      stringsAsFactors=FALSE
-    )
+                      stringsAsFactors=FALSE)
     
     result6 <- get_frontier(test6)
-    expect_identical(
-      sort(result6),
+    expect_equivalent(
+      result6,
       c("Scenario 1",  "Scenario 4", "Scenario 7")
     )
     
@@ -106,13 +104,98 @@ test_that(
     test7<-data.frame(.strategy_names=c("Scenario 1","Scenario 2","Scenario 3","Scenario 4","Scenario 5","Scenario 6","Scenario 7"),
                       .cost=c(25000,25000,50000,58500,58500,64266,91744),
                       .effect=c(10.2497,10.25,11.3,11.6588,11.6588,11.7,12.1),
-                      stringsAsFactors=FALSE
-                      
-    )
+                      stringsAsFactors=FALSE)
     result7 <- get_frontier(test7)
-    expect_identical(
-      sort(result7),
+    expect_equivalent(
+      result7,
       c("Scenario 2", "Scenario 4","Scenario 5", "Scenario 7")
+    )
+    
+    # Special Case:       Two Scenarios, both on frontier
+    # Frontier:           S1, S2
+    # Weakly Dominated:   None
+    # Strongly Dominated: None 
+    test8<-data.frame(.strategy_names=c("Scenario 1","Scenario 2"),
+                      .cost=c(26041,42000),
+                      .effect=c(10.2497,11.1654), stringsAsFactors=FALSE)
+    
+    result8<-get_frontier(test8)
+    expect_equivalent(
+      result8,
+      c("Scenario 1", "Scenario 2")
+    )
+    
+    
+    # Special Case:       Two Scenarios, first one is dominated
+    # Frontier:          S2
+    # Weakly Dominated:   None
+    # Strongly Dominated: S1 
+    test9<-data.frame(.strategy_names=c("Scenario 1","Scenario 2"),
+                      .cost=c(26041,21000),
+                      .effect=c(10.2497,11.1654), stringsAsFactors=FALSE)
+    
+    result9<-get_frontier(test9)
+    expect_equivalent(
+      result9,
+      c("Scenario 2")
+    )
+    
+    
+    # Special Case:       Three Scenarios, dominated by next neighbor
+    # Frontier:           S3
+    # Weakly Dominated:   None
+    # Strongly Dominated: S1, S2 
+    test10<-data.frame(.strategy_names=c("Scenario 1","Scenario 2","Scenario 3"),
+                       .cost=c(26041,21000,19000),
+                       .effect=c(10.2497,11.1654,11.5), stringsAsFactors=FALSE)
+    
+    result10<-get_frontier(test10)
+    expect_equivalent(
+      result10,
+      c("Scenario 3")
+    )
+    
+    
+    # Special Case:       Three Scenarios, one dominated by next neighbor
+    # Frontier:           S3
+    # Weakly Dominated:   None
+    # Strongly Dominated: S1, S2 
+    test11<-data.frame(.strategy_names=c("Scenario 1","Scenario 2","Scenario 3"),
+                       .cost=c(26041,27000,19000),
+                       .effect=c(10.2497,11.1654,11.5), stringsAsFactors=FALSE)
+    
+    result11<-get_frontier(test11)
+    expect_equivalent(
+      result11,
+      c("Scenario 3")
+    )
+    
+    # Special Case:       Just one strategy
+    # Frontier:           S1
+    # Weakly Dominated:   None
+    # Strongly Dominated: None
+    test12<-data.frame(.strategy_names=c("Scenario 1"),
+                       .cost=c(26041),
+                       .effect=c(10.2497), stringsAsFactors=FALSE)
+    
+    result12<-get_frontier(test12)
+    expect_equivalent(
+      result12,
+      c("Scenario 1")
+    )
+    
+    # Special Case:       Three Scenarios, one dominated with same effectiveness
+    # Frontier:           S2, S3
+    # Weakly Dominated:   None
+    # Strongly Dominated: S1 
+    test13 <- data.frame(.strategy_names=c("Scenario 1","Scenario 2","Scenario 3"),
+                         .cost = c(15, 10, 15),
+                         .effect = c(12, 12, 15), stringsAsFactors=FALSE)
+    
+    result13<-get_frontier(test13)
+    expect_equivalent(
+      result13,
+      c("Scenario 2","Scenario 3")
     )
   }
 )